Refactor into package and tests
